+++ ShufMath/Views/GamePlay/Main/Ui Elements/GameButtons/GameButtons.swift
@@ -33,26 +33,3 @@ struct GameButtons: View {
         .padding(10)
     }
 }
-
-struct ButtonStyleModifier: ViewModifier {
-    var paddingAmount: CGFloat = 8.0
-    var buttonText: String
-    var color: Color
-    
-    func body(content: Content) -> some View {
-        content
-            .buttonStyle(.plain)
-            .frame(maxWidth: .infinity)
-            .padding(paddingAmount)
-            .background(color)
-            .foregroundStyle(.white)
-            .cornerRadius(10)
-            .accessibilityLabel("Tap to \(buttonText)")
-    }
-}
-
-extension View {
-    func customButtonStyle(paddingAmount: CGFloat = 8.0, buttonText: String, color: Color) -> some View {
-        self.modifier(ButtonStyleModifier(paddingAmount: paddingAmount, buttonText: buttonText, color: color))
-    }
-}
